+++ releases/eucalyptus/3/bare/config/lms/docker_run_production.py
@@ -184,7 +184,7 @@ SESSION_COOKIE_HTTPONLY = config(
     "SESSION_COOKIE_HTTPONLY", default=True, formatter=bool
 )
 SESSION_COOKIE_SECURE = config(
-    "SESSION_COOKIE_SECURE", default=SESSION_COOKIE_SECURE, formatter=bool
+    "SESSION_COOKIE_SECURE", default=True, formatter=bool
 )
 SESSION_SAVE_EVERY_REQUEST = config(
     "SESSION_SAVE_EVERY_REQUEST", default=SESSION_SAVE_EVERY_REQUEST, formatter=bool