+++ releases/hawthorn/1/bare/config/cms/docker_run_production.py
@@ -203,7 +203,7 @@ SESSION_ENGINE = config(
     "SESSION_ENGINE", default="django.contrib.sessions.backends.cache"
 )
 SESSION_COOKIE_SECURE = config(
-    "SESSION_COOKIE_SECURE", default=SESSION_COOKIE_SECURE, formatter=bool
+    "SESSION_COOKIE_SECURE", default=True, formatter=bool
 )
 SESSION_SAVE_EVERY_REQUEST = config(
     "SESSION_SAVE_EVERY_REQUEST", default=SESSION_SAVE_EVERY_REQUEST, formatter=bool